Refactor GitHub Authentication logic, seperating UX from core logic.
