+++ Cli-CredentialHelper/Cli-CredentialHelper.csproj
@@ -61,6 +61,7 @@
     <Compile Include="AuthorityType.cs" />
     <Compile Include="Installer.cs" />
     <Compile Include="Interactivity.cs" />
+    <Compile Include="NativeMethods.cs" />
     <Compile Include="OperationArguments.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
