+++ Cli-CredentialHelper/Program.cs
@@ -1,11 +1,15 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.ComponentModel;
 using System.Diagnostics;
 using System.IO;
 using System.Linq;
+using System.Runtime.InteropServices;
+using System.Text;
 using System.Threading.Tasks;
 using Microsoft.Alm.Authentication;
 using Microsoft.Alm.Git;
+using Microsoft.Win32.SafeHandles;
 
 namespace Microsoft.Alm.CredentialHelper
 {
@@ -489,6 +493,8 @@ namespace Microsoft.Alm.CredentialHelper
                         || GithubAuthentication.GetAuthentication(operationArguments.TargetUri,
                                                                   GithubCredentialScope,
                                                                   secrets,
+                                                                  GithubCredentialPrompt, 
+                                                                  GithubAuthCodePrompt,
                                                                   out authority))
                     {
                         // set the authority type based on the returned value
@@ -530,7 +536,10 @@ namespace Microsoft.Alm.CredentialHelper
                     Trace.WriteLine("   authority it GitHub");
 
                     // return a GitHub authenitcation object
-                    return authority ?? new GithubAuthentication(GithubCredentialScope, secrets);
+                    return authority ?? new GithubAuthentication(GithubCredentialScope, 
+                                                                 secrets, 
+                                                                 GithubCredentialPrompt, 
+                                                                 GithubAuthCodePrompt);
 
                 case AuthorityType.MicrosoftAccount:
                     Trace.WriteLine("   authority is Microsoft Live");
@@ -636,7 +645,7 @@ namespace Microsoft.Alm.CredentialHelper
         {
             //const string EventSource = "Git Credential Manager";
 
-            /*** commented out due to UAC issues which require a proper installer to work around ***/
+            ///*** commented out due to UAC issues which require a proper installer to work around ***/
 
             //Trace.WriteLine("Program::LogEvent");
 
@@ -698,6 +707,168 @@ namespace Microsoft.Alm.CredentialHelper
                     listener.WriteLine(String.Format("Microsoft {0} version {0}", Program.Title, Version.ToString(3)));
                 }
             }
+        }
+
+        private static bool GithubCredentialPrompt(Uri targetUri, out string username, out string password)
+        {
+            // ReadConsole 32768 fail, 32767 ok 
+            // @linquize [https://github.com/Microsoft/Git-Credential-Manager-for-Windows/commit/a62b9a19f430d038dcd85a610d97e5f763980f85]
+            const int BufferReadSize = 32 * 1024 - 7;
+
+            StringBuilder buffer = new StringBuilder(BufferReadSize);
+            uint read = 0;
+            uint written = 0;
+
+            username = null;
+            password = null;
+
+            NativeMethods.FileAccess fileAccessFlags = NativeMethods.FileAccess.GenericRead | NativeMethods.FileAccess.GenericWrite;
+            NativeMethods.FileAttributes fileAttributes = NativeMethods.FileAttributes.Normal;
+            NativeMethods.FileCreationDisposition fileCreationDisposition = NativeMethods.FileCreationDisposition.OpenExisting;
+            NativeMethods.FileShare fileShareFlags = NativeMethods.FileShare.Read | NativeMethods.FileShare.Write;
+
+            using (SafeFileHandle stdout = NativeMethods.CreateFile(NativeMethods.ConsoleOutName, fileAccessFlags, fileShareFlags, IntPtr.Zero, fileCreationDisposition, fileAttributes, IntPtr.Zero))
+            using (SafeFileHandle stdin = NativeMethods.CreateFile(NativeMethods.ConsoleInName, fileAccessFlags, fileShareFlags, IntPtr.Zero, fileCreationDisposition, fileAttributes, IntPtr.Zero))
+            {
+                // read the current console mode
+                NativeMethods.ConsoleMode consoleMode;
+                if (!NativeMethods.GetConsoleMode(stdin, out consoleMode))
+                {
+                    int error = Marshal.GetLastWin32Error();
+                    throw new Win32Exception(error, "Unable to determine console mode (" + error + ").");
+                }
+
+                // instruct the user as to what they are expected to do
+                buffer.Append("Please enter your GitHub credentials for ")
+                      .Append(targetUri.Scheme)
+                      .Append("://")
+                      .Append(targetUri.DnsSafeHost)
+                      .Append(targetUri.PathAndQuery)
+                      .AppendLine();
+                if (!NativeMethods.WriteConsole(stdout, buffer, (uint)buffer.Length, out written, IntPtr.Zero))
+                {
+                    int error = Marshal.GetLastWin32Error();
+                    throw new Win32Exception(error, "Unable to write to standard output (" + error + ").");
+                }
+
+                // clear the buffer for the next operation
+                buffer.Clear();
+
+                // prompt the user for the username wanted
+                buffer.Append("username: ");
+                if (!NativeMethods.WriteConsole(stdout, buffer, (uint)buffer.Length, out written, IntPtr.Zero))
+                {
+                    int error = Marshal.GetLastWin32Error();
+                    throw new Win32Exception(error, "Unable to write to standard output (" + error + ").");
+                }
+
+                // clear the buffer for the next operation
+                buffer.Clear();
+
+                // read input from the user
+                if (!NativeMethods.ReadConsole(stdin, buffer, BufferReadSize, out read, IntPtr.Zero))
+                {
+                    int error = Marshal.GetLastWin32Error();
+                    throw new Win32Exception(error, "Unable to read from standard input (" + error + ").");
+                }
+
+                // record input from the user into local storage, stripping any eol chars
+                username = buffer.ToString(0, (int)read);
+                username = username.Trim(Environment.NewLine.ToCharArray());
+
+                // clear the buffer for the next operation
+                buffer.Clear();
+
+                // set the console mode to current without echo input
+                NativeMethods.ConsoleMode consoleMode2 = consoleMode ^ NativeMethods.ConsoleMode.EchoInput;
+
+                try
+                {
+                    if (!NativeMethods.SetConsoleMode(stdin, consoleMode2))
+                    {
+                        int error = Marshal.GetLastWin32Error();
+                        throw new Win32Exception(error, "Unable to set console mode (" + error + ").");
+                    }
+
+                    // prompt the user for password
+                    buffer.Append("password: ");
+                    if (!NativeMethods.WriteConsole(stdout, buffer, (uint)buffer.Length, out written, IntPtr.Zero))
+                    {
+                        int error = Marshal.GetLastWin32Error();
+                        throw new Win32Exception(error, "Unable to write to standard output (" + error + ").");
+                    }
+
+                    // clear the buffer for the next operation
+                    buffer.Clear();
+
+                    // read input from the user
+                    if (!NativeMethods.ReadConsole(stdin, buffer, BufferReadSize, out read, IntPtr.Zero))
+                    {
+                        int error = Marshal.GetLastWin32Error();
+                        throw new Win32Exception(error, "Unable to read from standard input (" + error + ").");
+                    }
+
+                    // record input from the user into local storage, stripping any eol chars
+                    password = buffer.ToString(0, (int)read);
+                    password = password.Trim(Environment.NewLine.ToCharArray());
+                }
+                catch { throw; }
+                finally
+                {
+                    // restore the console mode to its original value
+                    if (!NativeMethods.SetConsoleMode(stdin, consoleMode))
+                    {
+                        int error = Marshal.GetLastWin32Error();
+                        throw new Win32Exception(error, "Unable to set console mode (" + error + ").");
+                    }
+                }
+            }
+
+            return username != null
+                && password != null;
+        }
+
+        private static bool GithubAuthCodePrompt(Uri targetUri, GithubAuthenticationResultType resultType, out string authenticationCode)
+        {
+            // ReadConsole 32768 fail, 32767 ok 
+            // @linquize [https://github.com/Microsoft/Git-Credential-Manager-for-Windows/commit/a62b9a19f430d038dcd85a610d97e5f763980f85]
+            const int BufferReadSize = 32 * 1024 - 7;
+
+            StringBuilder buffer = new StringBuilder(BufferReadSize);
+            uint read = 0;
+            uint written = 0;
+
+            authenticationCode = null;
+
+            NativeMethods.FileAccess fileAccessFlags = NativeMethods.FileAccess.GenericRead | NativeMethods.FileAccess.GenericWrite;
+            NativeMethods.FileAttributes fileAttributes = NativeMethods.FileAttributes.Normal;
+            NativeMethods.FileCreationDisposition fileCreationDisposition = NativeMethods.FileCreationDisposition.OpenExisting;
+            NativeMethods.FileShare fileShareFlags = NativeMethods.FileShare.Read | NativeMethods.FileShare.Write;
+
+            using (SafeFileHandle stdout = NativeMethods.CreateFile(NativeMethods.ConsoleOutName, fileAccessFlags, fileShareFlags, IntPtr.Zero, fileCreationDisposition, fileAttributes, IntPtr.Zero))
+            using (SafeFileHandle stdin = NativeMethods.CreateFile(NativeMethods.ConsoleInName, fileAccessFlags, fileShareFlags, IntPtr.Zero, fileCreationDisposition, fileAttributes, IntPtr.Zero))
+            {
+                buffer.AppendLine()
+                      .Append("authcode: ");
+                if (!NativeMethods.WriteConsole(stdout, buffer, (uint)buffer.Length, out written, IntPtr.Zero))
+                {
+                    int error = Marshal.GetLastWin32Error();
+                    throw new Win32Exception(error, "Unable to write to standard output (" + error + ").");
+                }
+                buffer.Clear();
+
+                // read input from the user
+                if (!NativeMethods.ReadConsole(stdin, buffer, BufferReadSize, out read, IntPtr.Zero))
+                {
+                    int error = Marshal.GetLastWin32Error();
+                    throw new Win32Exception(error, "Unable to read from standard input (" + error + ").");
+                }
+
+                authenticationCode = buffer.ToString(0, (int)read);
+                authenticationCode = authenticationCode.Trim(Environment.NewLine.ToCharArray());
+            }
+
+            return authenticationCode != null;
         }
 
         [Conditional("DEBUG")]
